+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add get morphoelectric (me) model tool
+
 ## [0.1.0] - 19.09.2024
 
 ### Added